+++ layout21tetris/Cargo.toml
@@ -6,13 +6,15 @@ authors = ["Dan Fritchman <dan@fritch.mn>"]
 edition = "2018"
 license = "BSD-3-Clause"
 repository = "https://github.com/dan-fritchman/Layout21"
-version = "0.2.0"
+version = "0.2.1"
 workspace = "../"
+exclude = ["resources",]
 
 [dependencies]
 # Local workspace dependencies
-layout21raw = {path = "../layout21raw"}
-layout21utils = {path = "../layout21utils"}
+layout21raw = {path = "../layout21raw", version = "0.2.1"}
+layout21utils = {path = "../layout21utils", version = "0.2.1"}
+
 # Crates.io 
 derive_builder = "0.9"
 derive_more = "0.99.16"
