+++ layout21tetris/src/lib.rs
@@ -10,13 +10,15 @@ extern crate derive_builder;
 pub mod abs;
 pub mod bbox;
 pub mod cell;
+pub mod conv;
 pub mod coords;
+pub mod instance;
 pub mod interface;
+pub mod layout;
 pub mod library;
 pub mod outline;
 pub mod placement;
 pub mod placer;
-pub mod rawconv;
 pub mod stack;
 pub mod tracks;
 pub mod validate;
